+++ build_many_linux_32bit.sh
@@ -2,35 +2,13 @@
 set -e
 
 # To execute this script:
-# docker run --rm -v $PWD:/data local/manylinux32 /data/build_many_linux_32bit.sh
+# docker run --rm -v $PWD:/data local/manylinux_crypto32 /data/build_many_linux_32bit.sh
 
-export UAMQP_VERSION="0.1.0b4"
+export UAMQP_VERSION="0.1.0b5"
 
-export CPATH="/etc/ssl/include"
-export LIBRARY_PATH="/etc/ssl/lib"
-export LD_LIBRARY_PATH="/etc/ssl/lib"
-export OPENSSL_ROOT_DIR="/etc/ssl"
-
-# Build libuuid
-# pushd /tmp
-# curl -O https://cdn.kernel.org/pub/linux/utils/util-linux/v2.27/util-linux-2.27.1.tar.gz
-# tar xvf util-linux-2.27.1.tar.gz
-# cd util-linux-2.27.1
-# ./configure --disable-shared --disable-all-programs --enable-libuuid CFLAGS=-fPIC
-# make
-# make install
-# popd
-
-# Build OpenSSL
-pushd /tmp
-cd openssl-1.0.2n
-./config shared --openssldir=/etc/ssl
-make depend
-make
-make install
-cp /etc/ssl/lib/*.a /usr/local/lib/
-cp /etc/ssl/lib/*.so.* /usr/local/lib/
-popd
+export CPATH="/opt/pyca/cryptography/openssl/include"
+export LIBRARY_PATH="/opt/pyca/cryptography/openssl/lib"
+export OPENSSL_ROOT_DIR="/opt/pyca/cryptography/openssl"
 
 # Make sure Cython and Wheel are available in all env
 /opt/python/cp34-cp34m/bin/python -m pip install cython==0.27.3 wheel